+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## 1.1.13 (May 10, 2017)
+- Improvement: added pseudoLocaleName option
+- Issue: generate unnecessary `null` in `{pctBlack, number, null}` for pseudo text.
+
+## 1.1.12 (May 9, 2017)
+- Improvement: browser support
+
 ## 1.1.11 (May 6, 2017)
 - Improvement: leverages intl-messageformat-parser for improving pseudo generator